Move installation settings into own subsection
